atc: behaviour: across plan is flat list
Rather than having nested across plans to represent an `across` with
multiple vars, a single across step corresponds with a single across
plan. Changing the structure of the plan will make it easier to display
as a flat list in the web UI.

I'm a bit concerned about the lack of test coverage for the across step
builder, since it's kind of tricky logic, and the logic is coupled to
the ordering of steps in the plan...maybe worth making this less
coupled.

Signed-off-by: Aidan Oldershaw <aoldershaw@pivotal.io>